refactor: move routers and controllers into src/api folder #591
