+++ server/src/api/routes/assets.ts
@@ -1,8 +1,8 @@
 import express from "express";
 const router = express.Router();
-import { isLoggedIn } from "../middleware/auth";
+import { isLoggedIn } from "../../../middleware/auth";
 
-import * as assets from "@controllers/assets";
+import * as assets from "src/api/controllers/assets";
 
 router.put("/", isLoggedIn, assets.updateAll);
 router.get("/", isLoggedIn, assets.find);
