+++ server/src/api/routes/categories.ts
@@ -1,8 +1,8 @@
 import express from "express";
 const router = express.Router();
-import { isLoggedIn } from "../middleware/auth";
+import { isLoggedIn } from "../../../middleware/auth";
 
-import * as categories from "@controllers/categories";
+import * as categories from "src/api/controllers/categories";
 
 router.put("/", isLoggedIn, categories.updateV2);
