+++ server/src/api/routes/paymentMethods.ts
@@ -1,8 +1,8 @@
 import express from "express";
 const router = express.Router();
-import { isLoggedIn } from "../middleware/auth";
+import { isLoggedIn } from "../../../middleware/auth";
 
-import * as paymentMethods from "@controllers/paymentMethods";
+import * as paymentMethods from "src/api/controllers/paymentMethods";
 
 router.put("/", isLoggedIn, paymentMethods.update);
 router.get("/", isLoggedIn, paymentMethods.find);
