+++ server/src/api/routes/settings.ts
@@ -1,8 +1,8 @@
 import express from "express";
 const router = express.Router();
-import { isLoggedIn } from "../middleware/auth";
+import { isLoggedIn } from "../../../middleware/auth";
 
-import * as settings from "@controllers/settings";
+import * as settings from "src/api/controllers/settings";
 
 router.get("/", isLoggedIn, settings.find);
 router.patch("/", isLoggedIn, settings.update);
