+++ server/src/api/routes/transactions.ts
@@ -1,7 +1,7 @@
 import express from "express";
 const router = express.Router();
-import { isLoggedIn } from "../middleware/auth";
-import * as transactions from "@controllers/transactions";
+import { isLoggedIn } from "../../../middleware/auth";
+import * as transactions from "src/api/controllers/transactions";
 
 router.post("/", isLoggedIn, transactions.create);
