+++ server/src/api/routes/users.ts
@@ -1,12 +1,12 @@
 import express from "express";
 const router = express.Router();
-import * as users from "@controllers/users";
+import * as users from "src/api/controllers/users";
 import {
   isLoggedIn,
   isNotLoggedIn,
   forceNotLoggedIn,
   isAdmin,
-} from "../middleware/auth";
+} from "../../../middleware/auth";
 
 router.put("/", isLoggedIn, users.updateFields);
